Renamed A500 model file names
Closes #59
