+++ src/models/a500TimeSeriesPoint.ts
